Reduce max-lines ESLint rule from 1800 to 1200 lines
- Update package.json max-lines configuration from 1800 to 1200
- Split TimetableGrid.test.ts (2096 lines) into 6 focused test files:
  - TimetableGrid.states.test.ts: Basic states and rendering tests
  - TimetableGrid.navigation.test.ts: Focus management and navigation tests
  - TimetableGrid.sorting.test.ts: Issue sorting functionality tests
  - TimetableGrid.favorites.test.ts: Favorites and aliases tests
  - TimetableGrid.attendance.test.ts: Attendance and delta calculation tests
  - TimetableGrid.groups.test.ts: Group totals and formatting tests
- Remove original large test file
- All tests maintain complete functionality and coverage

Resolves #81
