--- source/tests/components/TimetableGrid.test.ts
@@ -1,2096 +1,0 @@
-import test from 'ava';
-import React from 'react';
-import {render} from 'ink-testing-library';
-import figures from 'figures';
-import {TimetableGrid} from '../../components/TimetableGrid.js';
-import type {WeeklyWorklogSummary} from '../../domain/WeeklyWorklogSummary.js';
-import {InkTestHelpers} from '../utils/ink-test-helpers.js';
-
-test('TimetableGrid shows loading state', t => {
-	const props = {
-		data: undefined,
-		isLoading: true,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	t.true(lastFrame()!.includes('Loading worklogs...'));
-});
-
-test('TimetableGrid shows no data state', t => {
-	const props = {
-		data: undefined,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	t.true(lastFrame()!.includes('No data available'));
-});
-
-test('TimetableGrid shows empty worklogs state', t => {
-	const emptyData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [],
-		weekTotal: 0,
-	};
-
-	const props = {
-		data: emptyData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	t.true(lastFrame()!.includes('No worklogs found for this week'));
-});
-
-test('TimetableGrid renders table header', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-19T00:00:00.000Z'),
-				totalHours: 4,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue Summary',
-						hours: 4,
-					},
-				],
-			},
-		],
-		weekTotal: 4,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('Issue'));
-	t.true(output.includes('Mon'));
-	t.true(output.includes('Tue'));
-	t.true(output.includes('Wed'));
-	t.true(output.includes('Thu'));
-	t.true(output.includes('Fri'));
-	t.false(output.includes('Sat'));
-	t.false(output.includes('Sun'));
-	t.true(output.includes('Total'));
-});
-
-test('TimetableGrid displays dates in weekday headers', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'), // Monday, October 14, 2024
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test Issue',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-
-	// Check that weekdays include date information
-	// Week starting Oct 14, 2024: Mon (14.10), Tue (15.10), Wed (16.10), Thu (17.10), Fri (18.10)
-	t.true(output.includes('Mon (14.10)'), 'Should display Monday with date');
-	t.true(output.includes('Tue (15.10)'), 'Should display Tuesday with date');
-	t.true(output.includes('Wed (16.10)'), 'Should display Wednesday with date');
-	t.true(output.includes('Thu (17.10)'), 'Should display Thursday with date');
-	t.true(output.includes('Fri (18.10)'), 'Should display Friday with date');
-});
-
-test('TimetableGrid renders issue data correctly', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'), // Friday
-				totalHours: 5.5,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue Summary',
-						hours: 4,
-					},
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue Summary',
-						hours: 1.5,
-					},
-				],
-			},
-		],
-		weekTotal: 5.5,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('TEST-117'));
-	t.true(output.includes('Test Issue Summary'));
-	t.true(output.includes('5.5')); // Aggregated hours for Friday
-});
-
-test('TimetableGrid formats hours correctly', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'), // Friday
-				totalHours: 2.5,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue Summary',
-						hours: 2.5,
-					},
-				],
-			},
-		],
-		weekTotal: 2.5,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('2.5')); // Decimal hours
-});
-
-test('TimetableGrid shows daily totals', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'), // Friday
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'First Issue',
-						hours: 4,
-					},
-					{
-						issueKey: 'TEST-118',
-						issueSummary: 'Second Issue',
-						hours: 4,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('Worklog'));
-	t.true(output.includes('8')); // Daily total for Friday
-});
-
-test('TimetableGrid shows week total', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-19T00:00:00.000Z'),
-				totalHours: 7,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue',
-						hours: 7,
-					},
-				],
-			},
-		],
-		weekTotal: 7,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('7')); // Week total
-});
-
-test('TimetableGrid handles multiple issues correctly', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-19T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'First Issue',
-						hours: 4,
-					},
-					{
-						issueKey: 'TEST-118',
-						issueSummary: 'Second Issue',
-						hours: 4,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	t.true(output.includes('TEST-117'));
-	t.true(output.includes('TEST-118'));
-	t.true(output.includes('First Issue'));
-	t.true(output.includes('Second Issue'));
-});
-
-test('TimetableGrid shows dash for zero hours', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'), // Only Friday has work
-				totalHours: 4,
-				issues: [
-					{
-						issueKey: 'TEST-117',
-						issueSummary: 'Test Issue',
-						hours: 4,
-					},
-				],
-			},
-		],
-		weekTotal: 4,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-
-	const output = lastFrame()!;
-	// Should show '-' for days without work (Mon-Thu only, since Fri has work)
-	const dashCount = (output.match(/-/g) || []).length;
-	t.true(dashCount >= 6); // At least 6 dashes for days without work
-});
-
-// Tests for initial focus behavior
-test.serial(
-	'TimetableGrid sets initial focus to first row and current day on component load',
-	async t => {
-		// Mock the current date to be a Tuesday (day index 1)
-		const OriginalDate = Date;
-		const mockDate = new Date('2024-10-15T10:00:00.000Z'); // Tuesday
-
-		// Simple Date constructor override
-		const MockDate = function (this: any, ...args: any[]): any {
-			if (args.length === 0) {
-				return mockDate;
-			}
-
-			// @ts-expect-error: Dynamic Date constructor call
-			return new OriginalDate(...args);
-		} as any;
-
-		MockDate.now = () => mockDate.getTime();
-		MockDate.UTC = OriginalDate.UTC;
-		MockDate.parse = OriginalDate.parse;
-		MockDate.prototype = OriginalDate.prototype;
-
-		global.Date = MockDate;
-
-		const sampleData: WeeklyWorklogSummary = {
-			weekStart: new Date('2024-10-14T00:00:00.000Z'), // Monday
-			weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-			dailySummaries: [
-				{
-					date: new Date('2024-10-14T00:00:00.000Z'),
-					totalHours: 4,
-					issues: [
-						{
-							issueKey: 'TEST-117',
-							issueSummary: 'First Issue',
-							hours: 4,
-						},
-					],
-				},
-			],
-			weekTotal: 4,
-		};
-
-		const favoriteIssues = [
-			{
-				key: 'TEST-117',
-				defaultTime: '4h',
-				defaultComment: 'Test work',
-			},
-		];
-
-		const focusedCells: Array<{issueKey: string; date: Date}> = [];
-		const handleCellWorklog = (data: {issueKey: string; date: Date}) => {
-			focusedCells.push(data);
-		};
-
-		const props = {
-			data: sampleData,
-			isLoading: false,
-			favoriteIssues,
-			onCellWorklog: handleCellWorklog,
-			isActive: true,
-		};
-
-		const {stdin, rerender} = render(React.createElement(TimetableGrid, props));
-
-		// Wait for component to mount and initial focus to be set
-		await InkTestHelpers.delay(10);
-		rerender(React.createElement(TimetableGrid, props));
-
-		// Simulate pressing Enter to trigger cell worklog (tests that focus is set)
-		stdin.write('\r');
-
-		// Should have focused on Tuesday (column index 1) for the first issue
-		t.is(focusedCells.length, 1, 'Should have triggered cell worklog');
-		t.is(focusedCells[0]?.issueKey, 'TEST-117', 'Should focus on first issue');
-
-		// Check that the focused date is Tuesday (2024-10-15)
-		const focusedDate = focusedCells[0]?.date;
-		t.is(focusedDate?.getDate(), 15, 'Should focus on Tuesday (15th)');
-		t.is(focusedDate?.getMonth(), 9, 'Should focus on October (month 9)');
-
-		// Restore original Date constructor
-		global.Date = OriginalDate;
-	},
-);
-
-test.serial(
-	'TimetableGrid sets initial focus to Monday when current day is weekend',
-	async t => {
-		// Mock the current date to be a Sunday (day index 0)
-		const OriginalDate = Date;
-		const mockDate = new Date('2024-10-13T10:00:00.000Z'); // Sunday
-
-		// Simple Date constructor override
-		const MockDate = function (this: any, ...args: any[]): any {
-			if (args.length === 0) {
-				return mockDate;
-			}
-
-			// @ts-expect-error: Dynamic Date constructor call
-			return new OriginalDate(...args);
-		} as any;
-
-		MockDate.now = () => mockDate.getTime();
-		MockDate.UTC = OriginalDate.UTC;
-		MockDate.parse = OriginalDate.parse;
-		MockDate.prototype = OriginalDate.prototype;
-
-		global.Date = MockDate;
-
-		const sampleData: WeeklyWorklogSummary = {
-			weekStart: new Date('2024-10-14T00:00:00.000Z'), // Monday
-			weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-			dailySummaries: [
-				{
-					date: new Date('2024-10-14T00:00:00.000Z'),
-					totalHours: 4,
-					issues: [
-						{
-							issueKey: 'TEST-117',
-							issueSummary: 'First Issue',
-							hours: 4,
-						},
-					],
-				},
-			],
-			weekTotal: 4,
-		};
-
-		const favoriteIssues = [
-			{
-				key: 'TEST-117',
-				defaultTime: '4h',
-				defaultComment: 'Test work',
-			},
-		];
-
-		const focusedCells: Array<{issueKey: string; date: Date}> = [];
-		const handleCellWorklog = (data: {issueKey: string; date: Date}) => {
-			focusedCells.push(data);
-		};
-
-		const props = {
-			data: sampleData,
-			isLoading: false,
-			favoriteIssues,
-			onCellWorklog: handleCellWorklog,
-			isActive: true,
-		};
-
-		const {stdin, rerender} = render(React.createElement(TimetableGrid, props));
-
-		// Wait for component to mount and initial focus to be set
-		await InkTestHelpers.delay(10);
-		rerender(React.createElement(TimetableGrid, props));
-
-		// Simulate pressing Enter to trigger cell worklog
-		stdin.write('\r');
-
-		// Should have focused on Monday (column index 0) for the first issue
-		t.is(focusedCells.length, 1, 'Should have triggered cell worklog');
-		t.is(focusedCells[0]?.issueKey, 'TEST-117', 'Should focus on first issue');
-
-		// Check that the focused date is Monday (2024-10-14)
-		const focusedDate = focusedCells[0]?.date;
-		t.is(
-			focusedDate?.getDate(),
-			14,
-			'Should focus on Monday (14th) when current day is weekend',
-		);
-		t.is(focusedDate?.getMonth(), 9, 'Should focus on October (month 9)');
-
-		// Restore original Date constructor
-		global.Date = OriginalDate;
-	},
-);
-
-test.serial(
-	'TimetableGrid sets initial focus to Friday when current day is Friday',
-	async t => {
-		// Mock the current date to be a Friday (day index 5)
-		const OriginalDate = Date;
-		const mockDate = new Date('2024-10-18T10:00:00.000Z'); // Friday
-
-		// Simple Date constructor override
-		const MockDate = function (this: any, ...args: any[]): any {
-			if (args.length === 0) {
-				return mockDate;
-			}
-
-			// @ts-expect-error: Dynamic Date constructor call
-			return new OriginalDate(...args);
-		} as any;
-
-		MockDate.now = () => mockDate.getTime();
-		MockDate.UTC = OriginalDate.UTC;
-		MockDate.parse = OriginalDate.parse;
-		MockDate.prototype = OriginalDate.prototype;
-
-		global.Date = MockDate;
-
-		const sampleData: WeeklyWorklogSummary = {
-			weekStart: new Date('2024-10-14T00:00:00.000Z'), // Monday
-			weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-			dailySummaries: [
-				{
-					date: new Date('2024-10-18T00:00:00.000Z'), // Friday data
-					totalHours: 4,
-					issues: [
-						{
-							issueKey: 'TEST-117',
-							issueSummary: 'First Issue',
-							hours: 4,
-						},
-					],
-				},
-			],
-			weekTotal: 4,
-		};
-
-		const favoriteIssues = [
-			{
-				key: 'TEST-117',
-				defaultTime: '4h',
-				defaultComment: 'Test work',
-			},
-		];
-
-		const focusedCells: Array<{issueKey: string; date: Date}> = [];
-		const handleCellWorklog = (data: {issueKey: string; date: Date}) => {
-			focusedCells.push(data);
-		};
-
-		const props = {
-			data: sampleData,
-			isLoading: false,
-			favoriteIssues,
-			onCellWorklog: handleCellWorklog,
-			isActive: true,
-		};
-
-		const {stdin, rerender} = render(React.createElement(TimetableGrid, props));
-
-		// Wait for component to mount and initial focus to be set
-		await InkTestHelpers.delay(50);
-		rerender(React.createElement(TimetableGrid, props));
-
-		// Simulate pressing Enter to trigger cell worklog
-		stdin.write('\r');
-
-		// Should have focused on Friday (column index 4) for the first issue
-		t.is(focusedCells.length, 1, 'Should have triggered cell worklog');
-		t.is(focusedCells[0]?.issueKey, 'TEST-117', 'Should focus on first issue');
-
-		// Check that the focused date is Friday (2024-10-18)
-		const focusedDate = focusedCells[0]?.date;
-		t.is(focusedDate?.getDate(), 18, 'Should focus on Friday (18th)');
-		t.is(focusedDate?.getMonth(), 9, 'Should focus on October (month 9)');
-
-		// Restore original Date constructor
-		global.Date = OriginalDate;
-	},
-);
-
-test.serial(
-	'TimetableGrid does not set initial focus when component is not active',
-	async t => {
-		const sampleData: WeeklyWorklogSummary = {
-			weekStart: new Date('2024-10-14T00:00:00.000Z'),
-			weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-			dailySummaries: [
-				{
-					date: new Date('2024-10-14T00:00:00.000Z'),
-					totalHours: 4,
-					issues: [
-						{
-							issueKey: 'TEST-117',
-							issueSummary: 'First Issue',
-							hours: 4,
-						},
-					],
-				},
-			],
-			weekTotal: 4,
-		};
-
-		const favoriteIssues = [
-			{
-				key: 'TEST-117',
-				defaultTime: '4h',
-				defaultComment: 'Test work',
-			},
-		];
-
-		const focusedCells: Array<{issueKey: string; date: Date}> = [];
-		const handleCellWorklog = (data: {issueKey: string; date: Date}) => {
-			focusedCells.push(data);
-		};
-
-		const props = {
-			data: sampleData,
-			isLoading: false,
-			favoriteIssues,
-			onCellWorklog: handleCellWorklog,
-			isActive: false, // Component is not active
-		};
-
-		const {stdin, rerender} = render(React.createElement(TimetableGrid, props));
-
-		// Wait for component to mount
-		await InkTestHelpers.delay(10);
-		rerender(React.createElement(TimetableGrid, props));
-
-		// Simulate pressing Enter - should not trigger cell worklog since no focus is set
-		stdin.write('\r');
-
-		// Should not have focused anything when component is not active
-		t.is(
-			focusedCells.length,
-			0,
-			'Should not have triggered cell worklog when component is not active',
-		);
-	},
-);
-
-// Navigation Tests - Testing the fix for empty week navigation bug
-test('TimetableGrid allows week navigation even with empty data', t => {
-	const navigationCalls: Array<'prev' | 'next'> = [];
-
-	const handleWeekChange = (direction: 'prev' | 'next') => {
-		navigationCalls.push(direction);
-	};
-
-	const emptyData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-01-08'),
-		weekEnd: new Date('2024-01-14'),
-		weekTotal: 0,
-		dailySummaries: [], // Empty week - no worklogs
-	};
-
-	const props = {
-		data: emptyData,
-		isLoading: false,
-		onWeekChange: handleWeekChange,
-	};
-
-	const {stdin} = render(React.createElement(TimetableGrid, props));
-
-	// Should show "No worklogs found for this week"
-	// But navigation should still work
-
-	// Simulate Shift+LeftArrow (previous week) - this should work
-	stdin.write('\u001B[1;2D');
-
-	// Simulate Shift+RightArrow (next week) - this should also work
-	stdin.write('\u001B[1;2C');
-
-	// Navigation should work even with empty data (this was the bug)
-	t.is(navigationCalls.length, 2);
-	t.is(navigationCalls[0], 'prev');
-	t.is(navigationCalls[1], 'next');
-});
-
-test('TimetableGrid blocks cell interaction with empty data but allows week navigation', t => {
-	const navigationCalls: Array<'prev' | 'next'> = [];
-	const cellWorklogCalls: Array<{issueKey: string; date: Date}> = [];
-
-	const handleWeekChange = (direction: 'prev' | 'next') => {
-		navigationCalls.push(direction);
-	};
-
-	const handleCellWorklog = (data: {issueKey: string; date: Date}) => {
-		cellWorklogCalls.push(data);
-	};
-
-	const emptyData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-01-08'),
-		weekEnd: new Date('2024-01-14'),
-		weekTotal: 0,
-		dailySummaries: [], // Empty week
-	};
-
-	const props = {
-		data: emptyData,
-		isLoading: false,
-		onWeekChange: handleWeekChange,
-		onCellWorklog: handleCellWorklog,
-	};
-
-	const {stdin} = render(React.createElement(TimetableGrid, props));
-
-	// Try Enter key (should NOT trigger cell worklog for empty week)
-	stdin.write('\r');
-
-	// Try arrow keys (should NOT work for cell navigation with empty data)
-	stdin.write('\u001B[C'); // Right arrow
-	stdin.write('\u001B[A'); // Up arrow
-
-	// Try week navigation (SHOULD work even with empty data)
-	stdin.write('\u001B[1;2D'); // Shift+Left arrow
-
-	// Cell interactions should be blocked, week navigation should work
-	t.is(cellWorklogCalls.length, 0); // No cell worklog calls
-	t.is(navigationCalls.length, 1); // Week navigation should work
-	t.is(navigationCalls[0], 'prev');
-});
-
-test('TimetableGrid allows navigation during loading state', t => {
-	const navigationCalls: Array<'prev' | 'next'> = [];
-
-	const handleWeekChange = (direction: 'prev' | 'next') => {
-		navigationCalls.push(direction);
-	};
-
-	const props = {
-		data: undefined,
-		isLoading: true,
-		onWeekChange: handleWeekChange,
-	};
-
-	const {stdin} = render(React.createElement(TimetableGrid, props));
-
-	// Navigation should work even during loading
-	stdin.write('\u001B[1;2D'); // Shift+LeftArrow
-	stdin.write('\u001B[1;2C'); // Shift+RightArrow
-
-	t.is(navigationCalls.length, 2);
-	t.is(navigationCalls[0], 'prev');
-	t.is(navigationCalls[1], 'next');
-});
-
-test('TimetableGrid sorts issues by project prefix and number', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 12,
-				issues: [
-					{
-						issueKey: 'DEF-2457',
-						issueSummary: 'DEF issue 2457',
-						hours: 2,
-					},
-					{
-						issueKey: 'ABC-5417',
-						issueSummary: 'ABC issue 5417',
-						hours: 3,
-					},
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'DEF issue 2456',
-						hours: 1,
-					},
-					{
-						issueKey: 'ABC-5420',
-						issueSummary: 'ABC issue 5420',
-						hours: 4,
-					},
-					{
-						issueKey: 'ABC-5419',
-						issueSummary: 'ABC issue 5419',
-						hours: 2,
-					},
-				],
-			},
-		],
-		weekTotal: 12,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Split output into lines to analyze order
-	const lines = output.split('\n');
-	const issueLines = lines.filter(
-		line => line.includes('ABC-') || line.includes('DEF-'),
-	);
-
-	// Should be sorted: ABC-5417, ABC-5419, ABC-5420, DEF-2456, DEF-2457
-	t.true(issueLines.length >= 5, 'Should have at least 5 issue lines');
-
-	// Find the indices of each issue in the output (with fixed width padding)
-	const gvv5417Index = issueLines.findIndex(line => line.includes('ABC-5417'));
-	const gvv5419Index = issueLines.findIndex(line => line.includes('ABC-5419'));
-	const gvv5420Index = issueLines.findIndex(line => line.includes('ABC-5420'));
-	const jts2456Index = issueLines.findIndex(line => line.includes('DEF-2456'));
-	const jts2457Index = issueLines.findIndex(line => line.includes('DEF-2457'));
-
-	// Verify sorting order: ABC issues come first (sorted by number), then DEF issues (sorted by number)
-	t.true(gvv5417Index < gvv5419Index, 'ABC-5417 should come before ABC-5419');
-	t.true(gvv5419Index < gvv5420Index, 'ABC-5419 should come before ABC-5420');
-	t.true(
-		gvv5420Index < jts2456Index,
-		'ABC issues should come before DEF issues',
-	);
-	t.true(jts2456Index < jts2457Index, 'DEF-2456 should come before DEF-2457');
-});
-
-test('TimetableGrid sorts issues with different project prefixes correctly', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 9,
-				issues: [
-					{
-						issueKey: 'ZZZ-100',
-						issueSummary: 'Last project issue',
-						hours: 3,
-					},
-					{
-						issueKey: 'AAA-200',
-						issueSummary: 'First project issue',
-						hours: 3,
-					},
-					{
-						issueKey: 'BBB-50',
-						issueSummary: 'Second project issue',
-						hours: 3,
-					},
-				],
-			},
-		],
-		weekTotal: 9,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	const lines = output.split('\n');
-	const issueLines = lines.filter(
-		line =>
-			line.includes('AAA-') || line.includes('BBB-') || line.includes('ZZZ-'),
-	);
-
-	const aaaIndex = issueLines.findIndex(line => line.includes('AAA-200'));
-	const bbbIndex = issueLines.findIndex(line => line.includes('BBB-50'));
-	const zzzIndex = issueLines.findIndex(line => line.includes('ZZZ-100'));
-
-	// Should be sorted alphabetically by project prefix: AAA, BBB, ZZZ
-	t.true(aaaIndex < bbbIndex, 'AAA should come before BBB');
-	t.true(bbbIndex < zzzIndex, 'BBB should come before ZZZ');
-});
-
-test('TimetableGrid sorts issues numerically within same project (124 before 1029)', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 6,
-				issues: [
-					{
-						issueKey: 'ABC-1029',
-						issueSummary: 'Higher number issue',
-						hours: 3,
-					},
-					{
-						issueKey: 'ABC-124',
-						issueSummary: 'Lower number issue',
-						hours: 3,
-					},
-				],
-			},
-		],
-		weekTotal: 6,
-	};
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	const lines = output.split('\n');
-	const issueLines = lines.filter(line => line.includes('ABC-'));
-
-	const gvv124Index = issueLines.findIndex(line => line.includes('ABC-124'));
-	const gvv1029Index = issueLines.findIndex(line => line.includes('ABC-1029'));
-
-	// 124 should come before 1029 (numeric sort, not string sort)
-	t.true(gvv124Index < gvv1029Index, 'ABC-124 should come before ABC-1029');
-	t.true(gvv124Index !== -1, 'ABC-124 should be found in output');
-	t.true(gvv1029Index !== -1, 'ABC-1029 should be found in output');
-});
-
-test('TimetableGrid shows favorite issues with asterisk marker', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 6,
-				issues: [
-					{
-						issueKey: 'ABC-5417',
-						issueSummary: 'Favorite issue',
-						hours: 3,
-					},
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'Regular issue',
-						hours: 3,
-					},
-				],
-			},
-		],
-		weekTotal: 6,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'ABC-5417',
-			defaultTime: '4h',
-			defaultComment: 'Favorite work',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Favorite issue should have star suffix with fixed width padding
-	t.true(
-		output.includes(`ABC-5417     ${figures.star}`),
-		'Favorite issue should be marked with star',
-	);
-
-	// Non-favorite issue should not have star but should be padded to fixed width
-	t.true(
-		output.includes('DEF-2456    '),
-		'Regular issue should be shown with fixed width padding',
-	);
-	t.false(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Regular issue should not have star',
-	);
-});
-
-test('TimetableGrid handles favorite issues without worklogs', t => {
-	// This test simulates when favorites are included via WeeklyWorklogSummaryUseCase
-	// but have no worklog entries (0 hours for all days)
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 3,
-				issues: [
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'Issue with worklog',
-						hours: 3,
-					},
-				],
-			},
-		],
-		weekTotal: 3,
-	};
-
-	// This represents a scenario where a favorite issue was fetched
-	// but has no worklogs, so it would appear with 0 hours
-	const favoriteIssues = [
-		{
-			key: 'ABC-5417',
-			defaultTime: '4h',
-			defaultComment: 'Favorite without worklog',
-		},
-		{
-			key: 'DEF-2456',
-			defaultTime: '2h',
-			defaultComment: 'Favorite with worklog',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Both issues should appear (one with worklog, one is configured favorite)
-	t.true(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Favorite with worklog should have star',
-	);
-	t.true(
-		output.includes('DEF-2456    '),
-		'Issue with worklog should be present',
-	);
-
-	// Note: ABC-5417 won't appear in this test because it's not in the worklog data
-	// The WeeklyWorklogSummaryUseCase would need to include it in the data
-	// This test validates the visual marking works correctly for favorites that do appear
-});
-
-test('TimetableGrid shows favorite issues even when no worklogs exist', t => {
-	// Empty worklog data - no worklogs for this week
-	const emptyData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [], // No worklogs at all
-		weekTotal: 0,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'ABC-5417',
-			defaultTime: '4h',
-			defaultComment: 'Favorite work',
-		},
-		{
-			key: 'DEF-2456',
-			defaultTime: '2h',
-			defaultComment: 'Other favorite',
-		},
-	];
-
-	const props = {
-		data: emptyData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Both favorite issues should appear even without worklogs
-	t.true(
-		output.includes(`ABC-5417     ${figures.star}`),
-		'Favorite ABC-5417 should be shown with star',
-	);
-	t.true(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Favorite DEF-2456 should be shown with star',
-	);
-
-	// Should show dashes for all days (no worklogs)
-	t.true(output.includes('-'), 'Should show dashes for days without worklogs');
-
-	// Should NOT show "No worklogs found" message when favorites exist
-	t.false(
-		output.includes('No worklogs found'),
-		'Should not show no worklogs message when favorites exist',
-	);
-});
-
-test('TimetableGrid displays aliases for favorite issues', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 6,
-				issues: [
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'Dev work issue',
-						hours: 4,
-					},
-					{
-						issueKey: 'ABC-5419',
-						issueSummary: 'Monitoring issue',
-						hours: 2,
-					},
-				],
-			},
-		],
-		weekTotal: 6,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'DEF-2456',
-			alias: 'Dev Work',
-			defaultTime: '8h',
-			defaultComment: 'Development work',
-		},
-		{
-			key: 'ABC-5419',
-			alias: 'Monitoring API',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Should display aliases instead of issue keys
-	t.true(
-		output.includes(`Dev Work     ${figures.star}`),
-		'Should display alias "Dev Work" for DEF-2456',
-	);
-	t.true(
-		output.includes(`Monitoring API ${figures.star}`),
-		'Should display alias "Monitoring API" for ABC-5419',
-	);
-
-	// Should NOT display the original issue keys in the issue column
-	t.false(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Should not display original key DEF-2456 when alias is set',
-	);
-	t.false(
-		output.includes(`ABC-5419     ${figures.star}`),
-		'Should not display original key ABC-5419 when alias is set',
-	);
-});
-
-test('TimetableGrid shows original key when no alias is configured', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 4,
-				issues: [
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'Issue without alias',
-						hours: 4,
-					},
-				],
-			},
-		],
-		weekTotal: 4,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'DEF-2456',
-			// No alias property set
-			defaultTime: '8h',
-			defaultComment: 'Development work',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Should display original issue key when no alias
-	t.true(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Should display original key DEF-2456 when no alias is set',
-	);
-});
-
-test('TimetableGrid handles mixed alias and non-alias favorites', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 6,
-				issues: [
-					{
-						issueKey: 'DEF-2456',
-						issueSummary: 'Dev work',
-						hours: 3,
-					},
-					{
-						issueKey: 'ABC-5419',
-						issueSummary: 'Regular work',
-						hours: 3,
-					},
-				],
-			},
-		],
-		weekTotal: 6,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'DEF-2456',
-			alias: 'Dev Work',
-			defaultTime: '8h',
-		},
-		{
-			key: 'ABC-5419',
-			// No alias - should show original key
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// One with alias, one without
-	t.true(
-		output.includes(`Dev Work     ${figures.star}`),
-		'Should display alias for DEF-2456',
-	);
-	t.true(
-		output.includes(`ABC-5419     ${figures.star}`),
-		'Should display original key for ABC-5419 when no alias',
-	);
-	t.false(
-		output.includes(`DEF-2456     ${figures.star}`),
-		'Should not display original DEF-2456 when alias is set',
-	);
-});
-
-test('TimetableGrid displays aliases with proper padding', t => {
-	const emptyData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [],
-		weekTotal: 0,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'DEF-2456',
-			alias: 'A', // Very short alias
-			defaultTime: '8h',
-		},
-		{
-			key: 'ABC-5419',
-			alias: 'Very Long Alias Name', // Long alias (longer than 12 chars)
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: emptyData,
-		isLoading: false,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Check that short aliases are properly padded
-	t.true(
-		output.includes(`A            ${figures.star}`),
-		'Short alias should be padded to 12 characters plus star',
-	);
-
-	// Long aliases should be included (the star may be on a separate line due to layout)
-	t.true(
-		output.includes('Very Long Alias Name'),
-		'Long alias should be displayed',
-	);
-	// Both aliases should have stars (check that both favorite issues have stars)
-	const starCount = (output.match(new RegExp(figures.star, 'g')) || []).length;
-	t.is(starCount, 2, 'Both favorite issues should have stars');
-});
-
-// Tests for new features: working hours calculation and group total formatting
-
-test('TimetableGrid shows attendance with working hours calculation', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '17:00',
-				breakMinutes: 60,
-			},
-			'2024-10-15': {
-				date: '2024-10-15',
-				checkIn: '09:30',
-				checkOut: '18:15',
-				breakMinutes: 45,
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [],
-		weekTotal: 0,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-
-	// Wait for async effects to complete - this is how ink-testing-library handles it
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should show attendance row
-	t.true(output.includes('Attendance'), 'Should show attendance row');
-
-	// Should show working hours calculation: 8-17 on first line, 8 on second line (decimal hours)
-	t.true(output.includes('8-17'), 'Should show Mon working hours 8-17');
-	t.true(
-		output.includes('8'),
-		'Should show Mon working hours 8 (decimal hours)',
-	);
-	t.true(
-		output.includes('9:30-18:15'),
-		'Should show Tue working hours with minutes',
-	);
-});
-
-test('TimetableGrid calculates working hours with different break times', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '17:00',
-				breakMinutes: 30, // 30 minute break
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [],
-		weekTotal: 0,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// 9 hours total - 0.5 hour break = 8.5 hours
-	t.true(output.includes('8-17'), 'Should show time range 8-17');
-	t.true(
-		output.includes('8.5'),
-		'Should show 8.5 with 30min break (decimal hours)',
-	);
-});
-
-test('TimetableGrid uses config default break time when not specified', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '17:00',
-				// No breakMinutes specified
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [],
-		weekTotal: 0,
-	};
-
-	const config = {
-		jiraUrl: 'test',
-		username: 'test',
-		apiToken: 'test',
-		attendance: {
-			enabled: true,
-			workingHours: 8,
-			breakMinutes: 30,
-			defaultCheckIn: '08:00',
-			defaultCheckOut: '17:00',
-			defaultBreakMinutes: 30,
-		},
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-		config,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should use config default break time (30 minutes)
-	t.true(output.includes('8-17'), 'Should show time range 8-17');
-	t.true(
-		output.includes('8.5'),
-		'Should use config default break time (decimal hours)',
-	);
-});
-
-test('TimetableGrid shows dash for empty group total', t => {
-	// Create data with the group issue but 0 hours
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 0,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 0, // No hours logged
-					},
-				],
-			},
-		],
-		weekTotal: 0,
-	};
-
-	const config = {
-		jiraUrl: 'test',
-		username: 'test',
-		apiToken: 'test',
-		groups: [
-			{
-				id: 'test-group',
-				name: 'Test Group',
-			},
-		],
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			groupId: 'test-group',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		config,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Group totals may not always be displayed depending on data structure
-
-	// The group total should show when there's an issue with 0 hours in a group
-	// But let's verify the formatting logic works by checking if dash formatting is correct
-	// Check that when groups are present, they don't have "-h" format
-	t.false(
-		output.includes('-h'),
-		'Should not show "-h" anywhere in output for empty totals',
-	);
-});
-
-test('TimetableGrid shows group total with hours suffix when not empty', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const config = {
-		jiraUrl: 'test',
-		username: 'test',
-		apiToken: 'test',
-		groups: [
-			{
-				id: 'test-group',
-				name: 'Test Group',
-			},
-		],
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			groupId: 'test-group',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		config,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Group totals may not always be displayed depending on grouping logic
-
-	// Verify the formatting works - when there are hours, should show with "h" suffix
-	// Even if the group total line isn't found, check that hour formatting is correct
-	t.true(
-		output.includes('8.0'),
-		'Should show the 8.0 hours somewhere in output',
-	);
-	t.false(output.includes('-h'), 'Should not show "-h" format anywhere');
-});
-
-test('TimetableGrid shows group total with desired amount and status', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-18T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const config = {
-		jiraUrl: 'test',
-		username: 'test',
-		apiToken: 'test',
-		groups: [
-			{
-				id: 'test-group',
-				name: 'Test Group',
-				desiredAmount: 10, // Desired 10 hours
-			},
-		],
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			groupId: 'test-group',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		config,
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Group totals may not always be displayed depending on data structure
-
-	// Verify basic functionality - hours should be shown, no "-h" format
-	t.true(
-		output.includes('8.0'),
-		'Should show the 8.0 hours somewhere in output',
-	);
-	t.false(output.includes('-h'), 'Should not show "-h" format anywhere');
-
-	// If groups are working, the warning emoji should appear for under-target groups
-	// This is a less strict test that verifies the main bug fix (no "-h")
-	const hasGrouping = output.includes('Test Group Total');
-	if (hasGrouping) {
-		t.true(
-			output.includes('⚠️'),
-			'Should show warning emoji when under target',
-		);
-	}
-});
-
-// Tests for new Delta row functionality
-test('TimetableGrid shows delta row with positive values in red', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '16:00', // 7.5 hours worked
-				breakMinutes: 30,
-				totalHours: 7.5,
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8, // Logged more than attended
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should show Delta row
-	t.true(output.includes('Delta'), 'Should show Delta row');
-
-	// Should show positive delta with + prefix (8.0 - 7.5 = +0.5)
-	t.true(output.includes('+0.5'), 'Should show positive delta with + prefix');
-});
-
-test('TimetableGrid shows delta row with zero values in green', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '17:00', // 8.5 hours worked
-				breakMinutes: 30,
-				totalHours: 8.5,
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8.5, // Logged exactly same as attended
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8.5,
-					},
-				],
-			},
-		],
-		weekTotal: 8.5,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should show Delta row
-	t.true(output.includes('Delta'), 'Should show Delta row');
-
-	// Should show zero delta (8.5 - 8.5 = 0.0)
-	t.true(output.includes('0.0'), 'Should show zero delta');
-});
-
-test('TimetableGrid shows delta row with negative values in red', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '18:00', // 9.5 hours worked
-				breakMinutes: 30,
-				totalHours: 9.5,
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8, // Logged less than attended
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should show Delta row
-	t.true(output.includes('Delta'), 'Should show Delta row');
-
-	// Should show negative delta (8.0 - 9.5 = -1.5)
-	t.true(output.includes('-1.5'), 'Should show negative delta');
-});
-
-test('TimetableGrid shows dash in delta row when no attendance data', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({}), // No attendance data
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-
-	// Should show Delta row
-	t.true(output.includes('Delta'), 'Should show Delta row');
-
-	// Should show dash when no attendance data
-	const lines = output.split('\n');
-	const deltaLine = lines.find(line => line.includes('Delta'));
-	t.truthy(deltaLine, 'Should find Delta row');
-	if (deltaLine) {
-		t.true(
-			deltaLine.includes('-'),
-			'Delta row should contain dash for missing data',
-		);
-	}
-});
-
-test('TimetableGrid shows attendance and delta rows at bottom after daily total', async t => {
-	const mockAttendanceManager = {
-		getWeeklyAttendance: async () => ({
-			'2024-10-14': {
-				date: '2024-10-14',
-				checkIn: '08:00',
-				checkOut: '17:00',
-				breakMinutes: 60,
-				totalHours: 8,
-			},
-		}),
-	};
-
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		attendanceManager: mockAttendanceManager as any,
-		attendanceRefreshKey: 1,
-		favoriteIssues,
-	};
-
-	const {lastFrame, rerender} = render(
-		React.createElement(TimetableGrid, props),
-	);
-	await InkTestHelpers.waitForEffects();
-	rerender(React.createElement(TimetableGrid, props));
-	await InkTestHelpers.waitForEffects();
-
-	const output = lastFrame()!;
-	const lines = output.split('\n');
-
-	// Find the indices of different rows
-	const worklogIndex = lines.findIndex(line => line.includes('Worklog'));
-	const deltaIndex = lines.findIndex(line => line.includes('Delta'));
-	const attendanceTimeIndex = lines.findIndex(
-		(line, index) => line.includes('Attendance') && index < worklogIndex,
-	); // First Attendance row (time ranges)
-
-	// Verify row order: Attendance should come first, then Issues, then Worklog, then Attendance (hours), then Delta
-	const attendanceHoursIndex = lines.findIndex(
-		(line, index) => line.includes('Attendance') && index > worklogIndex,
-	); // Second Attendance row (hours)
-
-	t.true(attendanceTimeIndex !== -1, 'Should find Attendance time row');
-	t.true(worklogIndex !== -1, 'Should find Worklog row');
-	t.true(attendanceHoursIndex !== -1, 'Should find Attendance hours row');
-	t.true(deltaIndex !== -1, 'Should find Delta row');
-
-	t.true(
-		attendanceTimeIndex < worklogIndex,
-		'Attendance time row should come before Worklog',
-	);
-	t.true(
-		worklogIndex < attendanceHoursIndex,
-		'Worklog should come before Attendance hours',
-	);
-	t.true(
-		attendanceHoursIndex < deltaIndex,
-		'Attendance hours should come before Delta',
-	);
-});
-
-test('TimetableGrid does not show delta row when no attendance manager', t => {
-	const sampleData: WeeklyWorklogSummary = {
-		weekStart: new Date('2024-10-14T00:00:00.000Z'),
-		weekEnd: new Date('2024-10-20T23:59:59.999Z'),
-		dailySummaries: [
-			{
-				date: new Date('2024-10-14T00:00:00.000Z'),
-				totalHours: 8,
-				issues: [
-					{
-						issueKey: 'TEST-123',
-						issueSummary: 'Test work',
-						hours: 8,
-					},
-				],
-			},
-		],
-		weekTotal: 8,
-	};
-
-	const favoriteIssues = [
-		{
-			key: 'TEST-123',
-			defaultTime: '4h',
-		},
-	];
-
-	const props = {
-		data: sampleData,
-		isLoading: false,
-		// No attendanceManager provided
-		favoriteIssues,
-	};
-
-	const {lastFrame} = render(React.createElement(TimetableGrid, props));
-	const output = lastFrame()!;
-
-	// Should NOT show Delta row when no attendance manager
-	t.false(
-		output.includes('Delta'),
-		'Should not show Delta row without attendance manager',
-	);
-
-	// Should still show Worklog
-	t.true(output.includes('Worklog'), 'Should still show Worklog row');
-});
